+++ packages/core/admin/admin/src/hooks/index.js
@@ -2,6 +2,7 @@ export { default as useModels } from './useModels';
 export { default as useFetchPermissionsLayout } from './useFetchPermissionsLayout';
 export { default as useFetchPluginsFromMarketPlace } from './useFetchPluginsFromMarketPlace';
 export { default as useFetchRole } from './useFetchRole';
+export { default as useMenu } from './useMenu';
 export { default as useRolesList } from './useRolesList';
 export { default as useSettingsHeaderSearchContext } from './useSettingsHeaderSearchContext';
 export { default as useSettingsMenu } from './useSettingsMenu';
