+++ packages/core/admin/admin/src/hooks/useMenu/utils/checkPermissions.js
