+++ packages/core/admin/admin/src/hooks/useMenu/utils/getGeneralLinks.js
