+++ packages/core/admin/admin/src/hooks/useMenu/utils/getPluginSectionLinks.js
