+++ packages/core/admin/admin/src/hooks/useMenu/utils/tests/checkPermissions.test.js
