+++ packages/core/admin/admin/src/hooks/useMenu/utils/tests/getGeneralLinks.test.js
