+++ packages/core/admin/admin/src/reducers.js
@@ -1,4 +1,4 @@
-import menuReducer from './components/LeftMenu/reducer';
+import menuReducer from './hooks/useMenu/reducer';
 import rbacProviderReducer from './components/RBACProvider/reducer';
 
 const reducers = {
